chore(dot/sync): move syncTask to fullsync.go
